Creating solution and moving Class1 to the console project
