+++ EmployeeDiscount/EmployeeDiscount.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 16
+VisualStudioVersion = 16.0.31613.86
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EmployeeDiscount", "EmployeeDiscount\EmployeeDiscount.csproj", "{3FF706ED-6014-47F7-94A3-F7E64F8A8EAC}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{3FF706ED-6014-47F7-94A3-F7E64F8A8EAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{3FF706ED-6014-47F7-94A3-F7E64F8A8EAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{3FF706ED-6014-47F7-94A3-F7E64F8A8EAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{3FF706ED-6014-47F7-94A3-F7E64F8A8EAC}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {1556E2D6-01ED-4BAC-A50C-77C3D49AF343}
+	EndGlobalSection
+EndGlobal
